# Fix deep copy max recursion failure - pydantic 1.9.1