@koxudaxi 2022-08-11
@eyalmor-ent 
I have released the PR as 0.13.1
Thank you very much 🙇